no commit message

+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/BackStockManager/Mes_BackStockDetailEntity.cs
@@ -41,7 +41,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 数量
         /// </summary>
         [Column("B_QTY")]
-        public decimal? B_Qty { get; set; }
+        public double? B_Qty { get; set; }
         /// <summary>
         /// 批次
         /// </summary>
@@ -56,7 +56,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 单价 edit 2019年4月15日16:34:30
         /// </summary>
         [Column("B_PRICE")]
-        public decimal? B_Price { get; set; }
+        public double? B_Price { get; set; }
         #endregion
 
         #region 扩展操作
