+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/CompUseHead/Mes_CompUseDetailEntity.cs
@@ -41,7 +41,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 数量
         /// </summary>
         [Column("C_QTY")]
-        public decimal? C_Qty { get; set; }
+        public double? C_Qty { get; set; }
         /// <summary>
         /// 批次
         /// </summary>
@@ -51,7 +51,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 价格
         /// </summary>
         [Column("C_PRICE")]
-        public decimal? C_Price { get; set; }
+        public double? C_Price { get; set; }
         /// <summary>
         /// 备注
         /// </summary>
