+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/GoodsList/Mes_GoodsEntity.cs
@@ -51,7 +51,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 价格
         /// </summary>
         [Column("G_PRICE")]
-        public decimal? G_Price { get; set; }
+        public double? G_Price { get; set; }
         /// <summary>
         /// 单位
         /// </summary>
@@ -61,18 +61,18 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 单位重量
         /// </summary>
         [Column("G_UNITWEIGHT")]
-        public decimal? G_UnitWeight { get; set; }
+        public double? G_UnitWeight { get; set; }
         
         /// <summary>
         /// 上限预警比例
         /// </summary>
         [Column("G_SUPER")]
-        public decimal? G_Super { get; set; }
+        public double? G_Super { get; set; }
         /// <summary>
         /// 下限预警比例
         /// </summary>
         [Column("G_LOWER")]
-        public decimal? G_Lower { get; set; }
+        public double? G_Lower { get; set; }
         /// <summary>
         /// 添加人
         /// </summary>
@@ -112,7 +112,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 包装规格
         /// </summary>
         [Column("G_UNITQTY")]
-        public decimal? G_UnitQty { get; set; }   
+        public double? G_UnitQty { get; set; }   
         /// <summary>
         /// 包装单位
         /// </summary>
@@ -137,12 +137,12 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 销售税率
         /// </summary>
         [Column("G_Otax")]
-        public decimal? G_Otax { get; set; }
+        public double? G_Otax { get; set; }
         /// <summary>
         /// 购进税率
         /// </summary>
         [Column("G_ITAX")]
-        public decimal? G_Itax { get; set; }
+        public double? G_Itax { get; set; }
         #endregion
 
         #region 扩展操作
@@ -188,7 +188,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 最新价格
         /// </summary>
          [NotMapped]
-        public decimal? M_Price { get; set; }
+        public double? M_Price { get; set; }
         #endregion
     }
 }
