+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/InventorySeach/Mes_InventoryEntity.cs
@@ -58,12 +58,12 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 数量
         /// </summary>
         [Column("I_QTY")]
-        public decimal? I_Qty { get; set; }  
+        public double? I_Qty { get; set; }  
         /// <summary>
         /// 数量
         /// </summary>
         [NotMapped]
-        public decimal? Qty { get; set; }
+        public double? Qty { get; set; }
         /// <summary>
         /// 批次批次以入库时间（yyyymmdd)
         /// </summary>
@@ -99,12 +99,12 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 总金额
         /// </summary>
         [NotMapped]
-        public decimal? AllMoney { get; set; }
+        public double? AllMoney { get; set; }
         /// <summary>
         /// 单个金额
         /// </summary>
         [NotMapped]
-        public decimal? OneMoney { get; set; }
+        public double? OneMoney { get; set; }
         /// <summary>
         /// 班组
         /// </summary>
@@ -114,11 +114,11 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         public string I_TeamCode { get; set; }
 
         [NotMapped]
-        public decimal? I_Price { get; set; }
+        public double? I_Price { get; set; }
         [NotMapped]
-        public decimal? G_Super { get; set; }
+        public double? G_Super { get; set; }
         [NotMapped]
-        public decimal? G_Lower { get; set; }
+        public double? G_Lower { get; set; }
         [NotMapped]
         public string G_State { get; set; }
         [NotMapped]
@@ -128,12 +128,12 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         [NotMapped]
         public string GoodsState { get; set; }
         [NotMapped]
-        public decimal? Price { get; set; }
+        public double? Price { get; set; }
         /// <summary>
         /// 销售税率
         /// </summary>
         [NotMapped]
-        public decimal? I_Otax { get; set; }
+        public double? I_Otax { get; set; }
         [NotMapped]
         public string S_Kind { get; set; }
         [NotMapped]
