+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/Mes_OtherInDetail/Mes_OtherInDetailEntity.cs
@@ -47,7 +47,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// </summary>
         /// <returns></returns>
         [Column("O_QTY")]
-        public decimal? O_Qty { get; set; }
+        public double? O_Qty { get; set; }
         /// <summary>
         /// 批次
         /// </summary>
@@ -59,7 +59,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// </summary>
         /// <returns></returns>
         [Column("O_PRICE")]
-        public decimal? O_Price { get; set; }
+        public double? O_Price { get; set; }
         /// <summary>
         /// 备注
         /// </summary>
