+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/OutWorkShopManager/Mes_OutWorkShopDetailEntity.cs
@@ -41,7 +41,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 数量
         /// </summary>
         [Column("O_QTY")]
-        public decimal? O_Qty { get; set; }
+        public double? O_Qty { get; set; }
         /// <summary>
         /// 批次
         /// </summary>
@@ -56,7 +56,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 价格
         /// </summary>
         [Column("O_PRICE")]
-        public decimal? O_Price { get; set; }
+        public double? O_Price { get; set; }
 
         #endregion
