+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/PickingMater/Mes_CollarDetailEntity.cs
@@ -50,13 +50,13 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// </summary>
         /// <returns></returns>
         [Column("C_UNITQTY")]
-        public decimal?  C_UnitQty { get; set; }
+        public double?  C_UnitQty { get; set; }
         /// <summary>
         /// 包装数量
         /// </summary>
         /// <returns></returns>
         [Column("C_QTY2")]
-        public decimal? C_Qty2 { get; set; }
+        public double? C_Qty2 { get; set; }
         /// <summary>
         /// 供应商编码
         /// </summary>
@@ -86,7 +86,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 数量
         /// </summary>
         [Column("C_QTY")]
-        public decimal? C_Qty { get; set; }
+        public double? C_Qty { get; set; }
         /// <summary>
         /// 批次
         /// </summary>
@@ -101,7 +101,7 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 价格
         /// </summary>
         [Column("C_PRICE")]
-        public decimal? C_Price { get; set; }
+        public double? C_Price { get; set; }
         /// <summary>
         /// 班组
         /// </summary>
