+++ Source/BS/Ayma.Framework.Module/Ayma.Application.Module/Ayma.Application.TwoDevelopment/MesDev/PurchaseHead/Mes_PurchaseDetailEntity.cs
@@ -46,12 +46,12 @@ namespace Ayma.Application.TwoDevelopment.MesDev
         /// 数量
         /// </summary>
         [Column("P_QTY")]
-        public decimal? P_Qty { get; set; }
+        public double? P_Qty { get; set; }
         /// <summary>
         /// 价格
         /// </summary>
         [Column("P_PRICE")]
-        public decimal? P_Price { get; set; }
+        public double? P_Price { get; set; }
         /// <summary>
         /// 批次
         /// </summary>
